Moved files to Shared folder. Created second boy for scotsman guide ranking. Fixed Footer Col.
