+++ src/Components/Login.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import './Shared/HomePage.css'
+import 'bootstrap/dist/css/bootstrap.min.css'
+import {Container, Col, Row, Image} from 'react-bootstrap'
+
+import React from 'react'
+
+function Login() {
+  return (
+    <div className='body-container'>Login</div>
+  )
+}
+
+export default Login
